requirements.txt
astroid==2.7.3
botocore==1.21.40
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
decorator==4.4.2
Flask==2.0.1
google-api-core==2.0.1
google-auth==2.0.2
google-cloud-core==2.0.0
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
greenlet==1.1.1
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
moviepy==1.0.3
mysql-connector-python==8.0.26
numpy==1.21.2
Pillow==8.3.2
platformdirs==2.3.0
proglog==0.1.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.10.2
python-dateutil==2.8.2
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
tqdm==4.62.2
urllib3==1.26.6
uWSGI==2.0.19.1
Werkzeug==2.0.1
wrapt==1.12.1
